swagger: Move path parameters out of #/parameters/
Fixing a validation error on editor.swagger.io, which does not like our
previous spec that contained parameter references to #/parameters/.

Instead, path based parameters are all moved to the path definition
schema as inlined values.

Note that this swagger validation error looks to be a bug on the
swagger side, but our workaround of moving parameters to the definitions
is in line with their Petstore example definition.