api-docs: Fix issue where our validation isn't reflected in our docs
Previously we were allowing users to omit {start-loc|destination}
address if a valid lat/lng pair were provided (and vice versa).

Since we're not implementing geocoding/reverse geocoding in the near
future, these cases will be considered invalid for the time being.

This commit updates the API docs to reflect the stricter validation.