add beta docs
Adds a placeholder for the beta docs so we can add
the beta docs to the swagger deploy pipeline.

When we release the beta endpoints then we will move
the existing swagger.json to v1/swagger.json, move
the beta/swagger.json to swagger.json, and remove
the beta branch.